move solution checker functions from toysolver_check.hs to ToySolver.Internal.SolutionChecker
